asddasdas

+++ server/services/questions.service.js
@@ -8,26 +8,31 @@ class QuestionsService {
     async getQuestions(accessToken, diagnosticID) {
         const userData = tokenService.validateAccessToken(accessToken);
         if (userData && userData.role === 'admin') {
-            const questions = await db.models.questionModel.findAll({
+            let questions = await db.models.questionModel.findAll({
                 where: {
                     diagnosticID
                 }
-            }).map((v) => new QuestionDTO(v))
-            const response = await Promise.all(questions.map(async (question) => {
-                return {
-                  text: question.text,
-                  answers: await answersService.getAnswers(accessToken, question.questionID)
-                }
-              }));
+            })
+            let response = []
+            if(questions !== undefined){
+                questions = questions.map((v) => new QuestionDTO(v))
+                response = await Promise.all(questions.map(async (question) => {
+                    return {
+                    type: question.type,
+                    text: question.text,
+                    answers: await answersService.getAnswers(accessToken, question.questionID)
+                    }
+                }));
+            }
             return response
         }
-        return null
+        return []
     }
 
     async createQuestion(accessToken, diagnosticID, data) {
         const userData = tokenService.validateAccessToken(accessToken);
         if (userData && userData.role === 'admin') {
-            const answer = await db.models.questionModel.create({ type: 'text', text: data.text, diagnosticID })
+            const answer = await db.models.questionModel.create({ type: data.type, text: data.text, diagnosticID })
             const response = new QuestionDTO(answer)
             return response
         }
@@ -50,7 +55,9 @@ class QuestionsService {
 
     async deleteQuestionsByID(diagnosticID) {
         const questionsID = await this.getQuestionsIDByDiagnosticID(diagnosticID)
-        questionsID.foreach((id) => { answersService.deleteAnswersByID(id) })
+        if( questionsID.length !== 0 ) { 
+            questionsID.foreach((id) => { answersService.deleteAnswersByID(id) })
+        }
         db.models.questionModel.destroy({
             where: {
                 diagnosticID
